Progress on rendering logic and states
Rendering logic for switching tabs seems to be well. Also the application is also able to remember the states of the forms, whether they've been opened or closed. Next we're going to work on making a pure component that makes forms for us. Then we'll try to start experimenting with how we can use states to enter in form data, and have it render in the resume section, which will likely involve lifting state up on the form data.