Lifting states to update form input and display it
We've made it so the application can now updatethe information, for personal info form, in real time as the user fills out the form. Fixed issue with isOpen boolean being passed incorrectly. Also started working on functionality for clearing and loading the resume, though we've got it working for personal info form. Now we're going to work on styling the personal info section, and then after we'll work on making the education form and its functionality.