Progress on resume and education form section
Finished designing the header of the resume, and how to add 'sections' onto the resume such as 'education', 'professional experience', etc. Started working on designing the education section for the edit-panel, worked on how to show schools on sidebar, and rendering form fields. Then started working on designing the individual items the user is going to add onto the resume. Now we're preparing to start working on the complex functionality of the education form.