Lots of progress on education form
Good progress on education form such as the design/css, but also how we render it. It seems like submission is good, and we can at least add schools to it. Though there are still some bugs to be worked out with loading in example data, working out the issues with implementing the sidebar list, we need to do editing, and deleting as well, and test the rendering for that as well.