Fixed 'load example' and improved responsive design
Fixed the issues involving loading example data. We deep copied the data, which fixed it as now the school objects or items in our state arrays won't have the same references to the ones in formData.js. We also did some media queries in order to make the website more responsive and overally look better on mobile.