Added font selection, fixed rendering issue, finished project.
Lastly we added font selection feature to fill up the customization tab. We also fixed rendering issue in Resume.jsx because we forgot to copy properly the array states. And so basically this project should be finished. The next thing when we have time is to actually host this project on github pages or somewhere else.