Major improvements to app
Big overhaul on how app is structured. Improved code structure in App and EditPanel so it's more readable. More low-level stuff will be handled in the App, allowing us to create functions, then EditPanel will generally receive more high-level functions to interact with. As well as this EditPanel is a lot more readable. Created or modified reusable components for making parts of the forms, such as FormButtons and InputGroup. This leaves developers with the easy job of just making the forms. Added uuid for React best practices. Animated drop down for the item-form-sections (AddItem) components. Also used refs and effects to affect the DOM resume node for best practices. However, we still one small caveat we want to fix with setting the font class, update the colors, and perhaps add a dark mode to this.